+++ doc/Changelog.md
@@ -3,6 +3,8 @@ Changelog {#Changelog}
 
 # git master
 
+* [122](https://github.com/BlueBrain/Brion/pull/122):
+  Add loadNeuron for binary and map compartment reports
 * [121](https://github.com/BlueBrain/Brion/pull/121):
   Performance optimizations for map and binary compartment reports, fix
   sub target loading of map compartment reports